feat: login

+++ todolist/views.py
@@ -6,6 +6,9 @@ from .models import todo
 
 
 def __insert_todo(req):
+    if not req.user.is_authenticated:
+        return redirect("user-login")
+
     todo_name = req.POST.get("todo_name")
     todo_name = todo_name.strip()
     mysql_todo = todo.objects.filter(todo_name=todo_name)
@@ -21,6 +24,8 @@ def __insert_todo(req):
 
 # todo 조회 및 생성
 def selectTodolist(req):
+    if not req.user.is_authenticated:
+        return redirect("user-login")
     # insert
     if req.method == "POST":
         __insert_todo(req)
@@ -31,6 +36,8 @@ def selectTodolist(req):
 
 # todo 수정
 def updateTodo(req, todo_name):
+    if not req.user.is_authenticated:
+        return redirect("user-login")
     get_todo = todo.objects.get(todo_name=todo_name)
     get_todo.status = True
     get_todo.save()
@@ -39,6 +46,8 @@ def updateTodo(req, todo_name):
 
 # todo 삭제
 def deleteTodo(req, todo_name):
+    if not req.user.is_authenticated:
+        return redirect("user-login")
     get_todo = todo.objects.get(todo_name=todo_name)
     get_todo.delete()
     return redirect("select-todolist")
